Add exported ToggleActivity and allow its use when locked
Other apps can use this activity to toggle the state of the torch or set
a specific brightness. This activity also has showWhenLocked enabled,
which fixes the main regression from targeting API 35 where the first
toggling of the torch via the quick settings tile required unlocking the
device. This causes an animation bug in the lock screen where the
background turns black until the screen is turned off and back on, the
status bar is tapped, or the device is unlocked. This is not something
that can be fixed in PixelLight since ToggleActivity does not have a UI.

Users who want to toggle the torch via lock screen shortcuts can do so
by hijacking the QR code scanner action.

Fixes: #12

Suggested-By: @e-t-l
Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>